Fix remove transforms failing on Mono.
Running an xml transform using xdt:Transform='Remove' would fail
on Mono with a null reference exception.

The XmlElementContext TargetNode would be null in Mono resulting
in a null reference when the name of the node was logged. This
would only occur if the original xml node existed.

Microsoft.Web.XmlTransform.XmlNodeException: Object reference not set to an instance of an object ---> System.NullReferenceException: Object reference not set to an instance of an object
  at Microsoft.Web.XmlTransform.Remove.RemoveNode ()
  at Microsoft.Web.XmlTransform.Remove.Apply ()
  at Microsoft.Web.XmlTransform.Transform.ApplyOnce ()
  at Microsoft.Web.XmlTransform.Transform.Execute (Microsoft.Web.XmlTransform.XmlElementContext context, System.String argumentString)
  --- End of inner exception stack trace ---
  at Microsoft.Web.XmlTransform.XmlTransformationLogger.LogErrorFromException (System.Exception ex)
  at Microsoft.Web.XmlTransform.Transform.Execute (Microsoft.Web.XmlTransform.XmlElementContext context, System.String argumentString)
  at Microsoft.Web.XmlTransform.XmlTransformation.HandleElement (Microsoft.Web.XmlTransform.XmlElementContext context)